requirements.txt
Django>=3.0,<4.0
djangorestframework==3.13.1
markdown==3.3.6
django-filter==21.1
django-extensions==3.1.5
requests==2.27.1
django-cors-headers==3.11.0
